refactor : 타입 파일 이름 통일
