--- service/src/types/Authorization/request.ts
@@ -1,15 +1,0 @@
-export interface SendCodeRequestBody {
-  phoneNumber: string;
-}
-
-export interface ConfirmVerificationRequestBody extends SendCodeRequestBody {
-  verificationCode: string;
-}
-
-export interface LoginRequestBody {
-  name: string;
-  phoneNumber: string;
-  hasCodeVerified: boolean;
-  privacyConsent: boolean;
-  marketingConsent: boolean;
-}
