+++ service/src/types/Authorization/type.ts
@@ -1,3 +1,19 @@
+export interface SendCodeRequestBody {
+  phoneNumber: string;
+}
+
+export interface ConfirmVerificationRequestBody extends SendCodeRequestBody {
+  verificationCode: string;
+}
+
+export interface LoginRequestBody {
+  name: string;
+  phoneNumber: string;
+  hasCodeVerified: boolean;
+  privacyConsent: boolean;
+  marketingConsent: boolean;
+}
+
 export type CodeResponse = {
   isSuccess: boolean;
   code: string;
