+++ service/src/types/lottery/type.ts
